Fix compiler warnings in deque_sizeof()

+++ Modules/_collectionsmodule.c
@@ -1511,8 +1511,8 @@ deque_sizeof(dequeobject *deque, void *unused)
     size_t res = _PyObject_SIZE(Py_TYPE(deque));
     size_t blocks;
     blocks = (size_t)(deque->leftindex + Py_SIZE(deque) + BLOCKLEN - 1) / BLOCKLEN;
-    assert(deque->leftindex + Py_SIZE(deque) - 1 ==
-           (blocks - 1) * BLOCKLEN + deque->rightindex);
+    assert(((size_t)deque->leftindex + (size_t)Py_SIZE(deque) - 1) ==
+           ((blocks - 1) * BLOCKLEN + (size_t)deque->rightindex));
     res += blocks * sizeof(block);
     return PyLong_FromSize_t(res);
 }
